Testing out the FedoraApi class.
