+++ src/IFedoraClient.php
@@ -27,46 +27,8 @@ namespace Islandora\Chullo;
  * @license  http://www.gnu.org/licenses/gpl-3.0.en.html GPL
  * @link     http://www.islandora.ca
  */
-interface IFedoraClient
+interface IFedoraClient extends IFedoraApi
 {
-    /**
-     * Gets the Fedora base uri (e.g. http://localhost:8080/fcrepo/rest)
-     *
-     * @return string
-     */
-    public function getBaseUri();
-
-    /**
-     * Gets a Fedora resource.
-     *
-     * @param string    $uri            Resource URI
-     * @param array     $headers        HTTP Headers
-     * @param string    $transaction    Transaction id
-     *
-     * @return mixed    String or binary content if 200. Null if 304.
-     */
-    public function getResource($uri = "",
-                                $headers = [],
-                                $transaction = "");
-    /**
-     * Gets a Fedora resoure's headers.
-     *
-     * @param string    $uri            Resource URI
-     * @param string    $transaction    Transaction id
-     *
-     * @return array    Headers of a resource. 
-     */
-    public function getResourceHeaders($uri = "",
-                                       $transaction = "");
-    /**
-     * Gets information about the supported HTTP methods, etc., for a Fedora resource.
-     *
-     * @param string    $uri            Resource URI
-     *
-     * @return string   Options of a resource.
-     */
-    public function getResourceOptions($uri = "");
-
     /**
      * Gets RDF metadata from Fedora.
      *
@@ -81,40 +43,6 @@ interface IFedoraClient
                              $transaction = "");
 
     /**
-     * Creates a new resource in Fedora.
-     *
-     * @param string    $uri            Resource URI
-     * @param string    $content        String or binary content
-     * @param array     $headers        HTTP Headers
-     * @param string    $transaction    Transaction id
-     * @param string    $checksum       SHA-1 checksum
-     *
-     * @return string   Uri of newly created resource
-     */
-    public function createResource($uri = "",
-                                   $content = null,
-                                   $headers = [],
-                                   $transaction = "",
-                                   $checksum = "");
-
-    /**
-     * Saves a resource in Fedora.
-     *
-     * @param string    $uri            Resource URI
-     * @param string    $content        String or binary content
-     * @param array     $headers        HTTP Headers
-     * @param string    $transaction    Transaction id
-     * @param string    $checksum       SHA-1 checksum
-     *
-     * @return null
-     */
-    public function saveResource($uri,
-                                 $content = null,
-                                 $headers = [],
-                                 $transaction = "",
-                                 $checksum = "");
-
-    /**
      * Saves RDF in Fedora.
      *
      * @param string            $uri            Resource URI
@@ -126,88 +54,4 @@ interface IFedoraClient
     public function saveGraph($uri,
                               \EasyRdf_Graph $graph,
                               $transaction = "");
-
-    /**
-     * Modifies a resource using a SPARQL Update query.
-     *
-     * @param string    $uri            Resource URI
-     * @param string    $sparql         SPARQL Update query
-     * @param array     $headers        HTTP Headers
-     * @param string    $transaction    Transaction id
-     *
-     * @return null
-     */
-    public function modifyResource($uri,
-                                   $sparql = "",
-                                   $headers = [],
-                                   $transaction = "");
-
-    /**
-     * Issues a DELETE request to Fedora.
-     *
-     * @param string    $uri            Resource URI
-     * @param string    $transaction    Transaction id
-     *
-     * @return null
-     */
-    public function deleteResource($uri,
-                                   $transaction = "");
-    /**
-     * Issues a COPY request to Fedora.
-     *
-     * @param string    $uri            Resource URI
-     * @param array     $destination    Destination URI
-     * @param string    $transaction    Transaction id
-     *
-     * @return string
-     */
-    public function copyResource($uri,
-                                 $destination,
-                                 $transaction = "");
-    /**
-     * Issues a MOVE request to Fedora.
-     *
-     * @param string    $uri            Resource URI
-     * @param array     $destination    Destination URI
-     * @param string    $transaction    Transaction id
-     *
-     * @return string 
-     */
-    public function moveResource($uri,
-                                 $destination,
-                                 $transaction = "");
-
-    /**
-     * Creates a new transaction.
-     *
-     * @return string   Transaction id
-     */
-    public function createTransaction();
-
-    /**
-     * Extends a transaction.
-     *
-     * @param string    $id Transaction id
-     *
-     * @return string   Fedora response
-     */
-    public function extendTransaction($id);
-
-    /**
-     * Commits a transaction.
-     *
-     * @param string    $id Transaction id
-     *
-     * @return boolean
-     */
-    public function commitTransaction($id);
-
-    /**
-     * Rolls back a transaction.
-     *
-     * @param string    $id Transaction id
-     *
-     * @return boolean
-     */
-    public function rollbackTransaction($id);
 }
